+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 2017-08-25
 
+- restore window resize handling
 - some investigation re "Websocket handler failed with Processor actor" when exiting the server 
   while some client currently connected. TODO check https://stackoverflow.com/a/41359874/830737 
 - handle chart height and minWidth